# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberLearn LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login response (student auth or admin totp challenge)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-totp:
    post:
      operationId: verifyTotp
      tags: [auth]
      summary: Verify admin TOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpVerifyInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "401":
          description: Invalid code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "204":
          description: Password changed
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── COURSES ───────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all students (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: adminCreateUser
      tags: [admin]
      summary: Create student account (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/suspend:
    put:
      operationId: adminSuspendUser
      tags: [admin]
      summary: Toggle user suspend/active (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SuspendInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{id}/enroll:
    put:
      operationId: adminEnrollUser
      tags: [admin]
      summary: Enroll student in course (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminEnrollInput"
      responses:
        "200":
          description: Enrolled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /admin/users/{id}/progress:
    get:
      operationId: adminGetUserProgress
      tags: [admin]
      summary: View student progress (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserProgress"
...
  /admin/audit-log:
    get:
      operationId: adminGetAuditLog
      tags: [admin]
      summary: Audit log (admin)
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get platform settings (admin)
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update platform settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────────────────
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    LoginResult:
      type: object
      required: [step]
      properties:
        step:
          type: string
          enum: [done, totp_required]
        accessToken:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/UserProfile"
        tempToken:
          type: ["string", "null"]
    AuthSuccess:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
...
    TotpVerifyInput:
      type: object
      required: [tempToken, totpCode]
      properties:
        tempToken:
          type: string
        totpCode:
          type: string
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    UserProfile:
      type: object
      required: [id, username, role, avatarColor, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [ADMIN, STUDENT]
        avatarColor:
          type: string
        totpEnabled:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
...
    CourseProgress:
      type: object
      required: [courseId, percent, completedLessons, totalLessons]
      properties:
        courseId:
          type: integer
        percent:
          type: number
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        lessonProgress:
          type: array
          items:
            $ref: "#/components/schemas/LessonProgress"
    Enrollment:
      type: object
      required: [id, userId, courseId, enrolledAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        enrolledAt:
          type: string
        completedAt:
          type: ["string", "null"]

    # ─── MODULE SCHEMAS ───────────────────────────────────────────────────────
...
    LessonProgress:
      type: object
      required: [lessonId, completed]
      properties:
        lessonId:
          type: integer
        watchedSeconds:
          type: integer
        completed:
          type: boolean
        lastAccessed:
          type: ["string", "null"]
...
    AdminUser:
      type: object
      required: [id, username, role, isActive, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        lastLogin:
          type: ["string", "null"]
        enrolledCourses:
          type: integer
        badgesEarned:
          type: integer
        avatarColor:
          type: string
    AdminUserInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
    SuspendInput:
      type: object
      required: [isActive]
      properties:
        isActive:
          type: boolean
    AdminEnrollInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer
    AdminUserProgress:
      type: object
      required: [userId, courses]
      properties:
        userId:
          type: integer
        courses:
          type: array
          items:
            $ref: "#/components/schemas/CourseProgress"
...
    AuditLogEntry:
      type: object
      required: [id, action, timestamp]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        username:
          type: ["string", "null"]
        action:
          type: string
        detail:
          type: ["string", "null"]
        timestamp:
          type: string
...
    PlatformSettings:
      type: object
      required: [platformName, adminPanelEnabled, certSignatureName, certOrgName]
      properties:
        platformName:
          type: string
        logoUrl:
          type: ["string", "null"]
        adminPanelEnabled:
          type: boolean
        certSignatureName:
          type: string
        certOrgName:
          type: string
        maintenanceMode:
          type: boolean
    PlatformSettingsUpdate:
      type: object
      properties:
        platformName:
          type: string
        logoUrl:
          type: string
        adminPanelEnabled:
          type: boolean
        certSignatureName:
          type: string
        certOrgName:
          type: string
        maintenanceMode:
          type: boolean
...