# lib/api-zod/src/generated/types/adminUserInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CyberLearn LMS API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUserInput {
  username: string;
  password: string;
  email?: string;
}